Fixed aio input handling for streams. Note: with cat FILE | ./pitch_pro... it appears that ordering of aio_reads is not guaranteed, so reduced number of reads to two. For ./pitch_pro... <FILE, no such issue occurs

+++ src/io_engine/io_engine_aio.h
@@ -15,7 +15,7 @@ namespace pitchstream
         private:
         // file descriptor, size of read buffer and max number of ios in flight (must be divisible by 2)
         int input_fd;
-        int buffer_size;
+        off_t buffer_size;
         int num_ios_inflight;
     };
